Safari-609.1.11.1/LayoutTests/fast/dom/Document/adoptNode-null.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests that document.adoptNode(null) throws a TypeError");

shouldThrowErrorName("document.adoptNode(null)", "TypeError");
shouldThrowErrorName("document.adoptNode()", "TypeError");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
