Safari-609.1.11.1/LayoutTests/fast/dom/Document/adoptNode-null-expected.txt

Tests that document.adoptNode(null) throws a TypeError

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.adoptNode(null) threw exception TypeError: Argument 1 ('node') to Document.adoptNode must be an instance of Node.
PASS document.adoptNode() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
